+++ siem_integrations/clx_query_service/clxquery/urls.py
@@ -1,0 +1,4 @@
+from django.conf.urls import re_path
+from clxquery import views
+
+urlpatterns = [re_path("clxquery/$", views.ExecuteClxQuery.as_view())]